refactor: abstract compiler
