+++ projects/vite-plugin-typst/examples/typst-cli-error/index.typ
@@ -2,3 +2,4 @@
 = Hello World
 
 This is a single file example.
+#
