+++ projects/vite-plugin-typst/examples/typst-cli-error/vite.config.mjs
@@ -2,15 +2,5 @@ import { defineConfig } from 'vite';
 import { TypstPlugin } from '@myriaddreamin/vite-plugin-typst';
 
 export default defineConfig({
-  plugins: [TypstPlugin({
-    compiler: 'typst-cli',
-    documents: ['**/*.typ'],
-    onCompile: (input, project, ctx) => {
-      const res = project.compileHtml(input);
-      if (res.hasError()) {
-        res.printErrors();
-        process.exit(1);
-      }
-    }
-  })],
+  plugins: [TypstPlugin({ compiler: 'typst-cli' })],
 });
